backend/pyproject.toml
[project]
name = "trexa-backend"
version = "0.1.0"
description = "Backend for Trexa file intelligence system"
authors = [
    { name = "Sandeep Vigneshwar", email = "psv641035@gmail.com" }
]
dependencies = [
    "fastapi",
    "pandas",
    "uvicorn[standard]",
    "pydantic",
    "sentence-transformers",
    "chromadb",
    "python-dotenv",
    "PyMuPDF",
    "python-docx",
    "watchdog"
]
requires-python = ">=3.8"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
